requirements.txt
langchain-groq
streamlit
groq
faiss-cpu
PyPDF2
langchain_google_genai
langchain_community
python-dotenv
pypdf
langchain
docx2txt
bs4
openpyxl
requests